# api/openapi.yaml
openapi: 3.1.0
info:
  title: TallyKeep
  version: 0.1.0-dev
paths:
...
  /api/v1/unlock:
    post:
      tags:
      - unlock
      summary: Post Unlock
      operationId: post_unlock_api_v1_unlock_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlockResponse'
        '401':
          description: Bad passphrase
        '503':
          description: Crypto parameters not initialized — first run required
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/unlock/initialize:
    post:
      tags:
      - unlock
      summary: Post Unlock Initialize
      operationId: post_unlock_initialize_api_v1_unlock_initialize_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitializeResponse'
        '409':
          description: Already initialized
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InitializeResponse:
      properties:
        initialized:
          type: boolean
          title: Initialized
        unlocked:
          type: boolean
          title: Unlocked
      type: object
      required:
      - initialized
      - unlocked
      title: InitializeResponse
...
    UnlockRequest:
      properties:
        passphrase:
          type: string
          maxLength: 4096
          minLength: 1
          title: Passphrase
      type: object
      required:
      - passphrase
      title: UnlockRequest
    UnlockResponse:
      properties:
        unlocked:
          type: boolean
          title: Unlocked
      type: object
      required:
      - unlocked
      title: UnlockResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...